slabAlloc

+++ lib/windows/windows_info.odin
@@ -53,11 +53,13 @@ initWindowsInfo :: proc() {
 	systemInfo: coreWin.SYSTEM_INFO
 	coreWin.GetSystemInfo(&systemInfo)
 	windows_info.min_page_size = int(systemInfo.dwAllocationGranularity)
-	windows_info.min_page_size_mask = int(math.maskUpperBits(math.ctz(windows_info.min_page_size)))
+	windows_info.min_page_size_mask = int(
+		math.maskUpperBits(math.ctz(uint(windows_info.min_page_size))),
+	)
 	// NOTE: windows large pages require nonsense: https://stackoverflow.com/questions/42354504/enable-large-pages-in-windows-programmatically
 	windows_info.min_large_page_size = int(coreWin.GetLargePageMinimum())
 	windows_info.min_large_page_size_mask = int(
-		math.maskUpperBits(math.ctz(windows_info.min_large_page_size)),
+		math.maskUpperBits(math.ctz(uint(windows_info.min_large_page_size))),
 	)
 	windows_info.is_initialized = true
 }
